feat: extract SpectatorScreen feed cards into shared FeedCard component (closes #526)
- Move SpeechCard/WolfChatCard/SystemRow/AgentEpisodeCard/FeedItem and feed
  pure functions/constants to components/FeedCard.jsx
- Split feed-card CSS into FeedCard.module.css; trim SpectatorScreen.module.css
- Add lib/agentLinks.js (agentDetailPath/searchForViewerMode) shared by
  FeedCard and SpectatorScreen RightPane
- Drop useParams from cards; FeedItem receives sessionId via props (screen-agnostic)
- SpectatorScreen imports FeedItem/SystemRow; display/behavior unchanged
- Update doc/FrontendDesign.md (FeedCard component, §6.2 table)
